upgrade burn version

+++ src/rl_algorithm/preload_net/utils.rs
@@ -7,7 +7,7 @@ use ndarray::{Array1, Array2, Array3};
 use crate::rl_algorithm::base::rl_utils::{ndarray2tensor1, tensor2ndarray1};
 
 pub fn generate_flatten_idx<B: Backend>(
-    traj_length: Array1<i64>,
+    traj_length: Array1<i32>,
     max_traj_len: usize,
     device: &B::Device,
 ) -> Tensor<B, 1, Int> {
@@ -17,12 +17,12 @@ pub fn generate_flatten_idx<B: Backend>(
         .iter()
         .enumerate()
         .map(|(i, len)| {
-            let start = (i * max_traj_len) as i64;
-            let end = start as i64 + *len;
+            let start = (i * max_traj_len) as i32;
+            let end = start as i32 + *len;
             (start..end)
         })
         .flatten()
-        .collect::<Vec<i64>>();
+        .collect::<Vec<i32>>();
     let arr = Array1::from(vec);
     return ndarray2tensor1(arr, device);
 }
